requirements.txt
telethon
pandas
python-dotenv
requests
gspread
google-auth
